Avoid binding for avatarPickerMenuItem
