+++ src/test/java/com/faforever/client/fx/OnlinePlayerContextMenuControllerTest.java
@@ -5,6 +5,7 @@ import com.faforever.client.builders.AvatarBeanBuilder;
 import com.faforever.client.builders.GameBeanBuilder;
 import com.faforever.client.builders.PlayerBeanBuilder;
 import com.faforever.client.chat.InitiatePrivateChatEvent;
+import com.faforever.client.domain.AvatarBean;
 import com.faforever.client.domain.GameBean;
 import com.faforever.client.domain.PlayerBean;
 import com.faforever.client.game.JoinGameHelper;
@@ -22,6 +23,7 @@ import com.google.common.collect.Sets;
 import com.google.common.eventbus.EventBus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.testfx.util.WaitForAsyncUtils;
 
@@ -35,6 +37,7 @@ import static com.faforever.client.player.SocialStatus.FOE;
 import static com.faforever.client.player.SocialStatus.FRIEND;
 import static com.faforever.client.player.SocialStatus.OTHER;
 import static com.faforever.client.player.SocialStatus.SELF;
+import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -88,37 +91,13 @@ public class OnlinePlayerContextMenuControllerTest extends UITest {
         notificationService, playerService, replayService, uiService) {};
     loadFxml("theme/player_context_menu.fxml", clazz -> instance, instance);
 
-    player = PlayerBeanBuilder.create().defaultValues().username(TEST_USER_NAME).socialStatus(SELF).avatar(null).game(new GameBean()).get();
+    player = PlayerBeanBuilder.create().defaultValues().username(TEST_USER_NAME).socialStatus(OTHER).avatar(null).game(new GameBean()).get();
     instance.setPlayer(player);
     WaitForAsyncUtils.waitForFxEvents();
   }
 
   @Test
   public void testShowCorrectItems() {
-    assertThat(instance.showUserInfo.isVisible(), is(true));
-    assertThat(instance.sendPrivateMessageItem.isVisible(), is(false));
-    assertThat(instance.copyUsernameItem.isVisible(), is(true));
-    assertThat(instance.colorPickerMenuItem.isVisible(), is(false));
-    assertThat(instance.inviteItem.isVisible(), is(false));
-    assertThat(instance.addFriendItem.isVisible(), is(false));
-    assertThat(instance.removeFriendItem.isVisible(), is(false));
-    assertThat(instance.addFoeItem.isVisible(), is(false));
-    assertThat(instance.removeFoeItem.isVisible(), is(false));
-    assertThat(instance.reportItem.isVisible(), is(false));
-    assertThat(instance.joinGameItem.isVisible(), is(false));
-    assertThat(instance.watchGameItem.isVisible(), is(false));
-    assertThat(instance.viewReplaysItem.isVisible(), is(true));
-    assertThat(instance.kickGameItem.isVisible(), is(false));
-    assertThat(instance.kickLobbyItem.isVisible(), is(false));
-    assertThat(instance.broadcastMessage.isVisible(), is(false));
-    assertThat(instance.avatarPickerMenuItem.isVisible(), is(true));
-  }
-
-  @Test
-  public void testShowCorrectItemsForOther() {
-    player.setSocialStatus(OTHER);
-    WaitForAsyncUtils.waitForFxEvents();
-
     assertThat(instance.showUserInfo.isVisible(), is(true));
     assertThat(instance.sendPrivateMessageItem.isVisible(), is(true));
     assertThat(instance.copyUsernameItem.isVisible(), is(true));
@@ -136,6 +115,30 @@ public class OnlinePlayerContextMenuControllerTest extends UITest {
     assertThat(instance.kickLobbyItem.isVisible(), is(false));
     assertThat(instance.broadcastMessage.isVisible(), is(false));
     assertThat(instance.avatarPickerMenuItem.isVisible(), is(false));
+  }
+
+  @Test
+  public void testShowCorrectItemsForSelf() {
+    player.setSocialStatus(SELF);
+    WaitForAsyncUtils.waitForFxEvents();
+
+    assertThat(instance.showUserInfo.isVisible(), is(true));
+    assertThat(instance.sendPrivateMessageItem.isVisible(), is(false));
+    assertThat(instance.copyUsernameItem.isVisible(), is(true));
+    assertThat(instance.colorPickerMenuItem.isVisible(), is(false));
+    assertThat(instance.inviteItem.isVisible(), is(false));
+    assertThat(instance.addFriendItem.isVisible(), is(false));
+    assertThat(instance.removeFriendItem.isVisible(), is(false));
+    assertThat(instance.addFoeItem.isVisible(), is(false));
+    assertThat(instance.removeFoeItem.isVisible(), is(false));
+    assertThat(instance.reportItem.isVisible(), is(false));
+    assertThat(instance.joinGameItem.isVisible(), is(false));
+    assertThat(instance.watchGameItem.isVisible(), is(false));
+    assertThat(instance.viewReplaysItem.isVisible(), is(true));
+    assertThat(instance.kickGameItem.isVisible(), is(false));
+    assertThat(instance.kickLobbyItem.isVisible(), is(false));
+    assertThat(instance.broadcastMessage.isVisible(), is(false));
+    assertThat(instance.avatarPickerMenuItem.isVisible(), is(true));
   }
 
   @Test
@@ -290,5 +293,20 @@ public class OnlinePlayerContextMenuControllerTest extends UITest {
     assertTrue(instance.kickGameItem.isVisible());
     assertTrue(instance.kickLobbyItem.isVisible());
     assertTrue(instance.moderatorActionSeparator.isVisible());
+  }
+
+  @Test
+  public void testOnSelectAvatar() throws Exception {
+    player.setSocialStatus(SELF);
+    instance.avatarComboBox.show();
+
+    WaitForAsyncUtils.waitForAsyncFx(100_000, () -> instance.avatarComboBox.getSelectionModel().select(2));
+
+    ArgumentCaptor<AvatarBean> captor = ArgumentCaptor.forClass(AvatarBean.class);
+    verify(avatarService).changeAvatar(captor.capture());
+
+    AvatarBean avatarBean = captor.getValue();
+    assertThat(avatarBean.getUrl(), equalTo(new URL("http://www.example.com/avatar2.png")));
+    assertThat(avatarBean.getDescription(), is("Avatar Number #2"));
   }
 }
